requirements.txt
fastapi
uvicorn
websockets
mysql-connector-python
python-dotenv
pydantic
streamlit
requests
pandas